Added getters to character.py! Need to create a main.py and create a basic script that will ask the user to press a button to create a random npc.
I also need to do some tests. Because most of this is created randomly, it's pretty difficult to create unittests for it. I will do some research in case there's sometihng I'm missing, but I may just have to live with manual testing for this one even though I was hoping to build a testing suite for it.