After some debugging and brief testing, it appears to be working! I also added a set_skills_values function, and changed the self._skills to a dictionary to accomodate for skills values that I had somehow missed the first time around.
Skill values are given a random set of base stats, and then an added bonus based on education * 2, as in the rulebook.

I want to automatically add a random brawl skill, Sanity, Health, and possibly Luck to every character. I personally don't care about movement speed, so I probably won't add that. I may change my mind in the future, but it shouldn't be hard to add it should that be the case.