Added functionality for sanity, health, and luck! Also worked a bit on the visuals of the character generator below.
sanity is typically just POWER to begin with, but I wanted to add a little spice and variety to the NPCs, so I opted for a +/- 5 from starting POW.

Luck is determined by 3d6 * 5. There is a setter and getter for this.

Health is determined by size + constitution floor divided by 10. So if size is 40 and constitution is 45, then that's 85 // 10 = 8.5 = 8