Refactored the set_occupation so that it correctly read the occupations_list. Also changed self._occupation to set_occupation(), instead of setting it manually.
After a few initial tests, everything seems to work!!!!

I'm going to fiddle with it some more maybe later today or tomorrow, and then take a break from it and work on something else. I do need to also write a readme for it, however, so I'll need to take some time out to bang that out.